Dont need switch turn function
So, we don't really need the switch turn function. However, when i dont use it, it goes on for a certain amount of time until it gives an error. Also, the player and enemy stats are not changing after their turn. I am unsure as to why this is happening. Take a look please